refactor(assets): rename to kebab-case, doc types, sync cdk pins
Address review feedback from @Hweinstock:
- rename AssetManager.ts -> manager.ts (only non-React .ts in the repo
  with an uppercase name; feature dirs use role-named lowercase files
  like core/project/manager.tsx).
- document AssetFile vs EmbeddedFile so the distinction between a
  resolved tree entry and a raw Bun.embeddedFiles blob is explicit.
- sync src/assets/cdk/package.json to the minor-version pins landed in
  #1777 (aws-cdk-lib ~2.261.0, @aws/agentcore-cdk 0.1.0-alpha.45, etc.).
