+++ src/assetManager/index.ts
@@ -1,2 +1,2 @@
-export { AssetManager } from "./AssetManager";
+export { AssetManager } from "./manager";
 export { ASSET_NAMES, type AssetName, type AssetVariables } from "./types";
