+++ src/assetManager/manager.ts
